Add cluster functionality
