+++ internal/hypvsr/cluster.go
@@ -3,114 +3,148 @@ package hypvsr
 import (
 	"errors"
 	"os"
-	"path/filepath"
 
-	"github.com/enkodr/machina/internal/config"
+	"github.com/enkodr/machina/internal/osutil"
 )
 
 // Cluster holds the configuration details for a cluster of machines
 type Cluster struct {
-	config   *config.Config
+	baseDir  string
 	Kind     string    `yaml:"kind"`     // Kind of the resource, should be 'Cluster'
 	Name     string    `yaml:"name"`     // Name of the cluster. Must be unique in the system
 	Machines []Machine `yaml:"machines"` // List of machines in the cluster
 }
 
-func (c *Cluster) Prepare() error {
-	return nil
-}
-
 func (c *Cluster) CreateDir() error {
 	// Check if VM already exists
-	_, err := os.Stat(filepath.Join(c.config.Directories.Clusters, c.Name))
+	_, err := os.Stat(c.baseDir)
 	if !os.IsNotExist(err) {
 		return errors.New("cluster already exists")
 	}
+	osutil.MkDir(c.baseDir)
 
-	return nil
-}
-
-func (c *Cluster) Create() error {
 	for _, vm := range c.Machines {
-		vm.Create()
+		vm.baseDir = c.baseDir
+		err := vm.CreateDir()
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
 
-func (c *Cluster) Start() error {
+func (c *Cluster) Prepare() error {
+	// Run the prepare for each of the machines
 	for _, vm := range c.Machines {
-		vm.Start()
-	}
-	return nil
-}
-
-func (c *Cluster) Stop() error {
-	for _, vm := range c.Machines {
-		vm.Stop()
-	}
-	return nil
-}
-
-func (c *Cluster) ForceStop() error {
-	for _, vm := range c.Machines {
-		vm.ForceStop()
-	}
-	return nil
-}
-
-func (c *Cluster) Status() (string, error) {
-	for _, vm := range c.Machines {
-		vm.Status()
-	}
-	return "", nil
-}
-
-func (c *Cluster) Delete() error {
-	for _, vm := range c.Machines {
-		vm.Delete()
-	}
-	return nil
-}
-
-func (c *Cluster) CopyContent(origin string, dest string) error {
-	for _, vm := range c.Machines {
-		vm.CopyContent(origin, dest)
+		err := vm.Prepare()
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
 
 func (c *Cluster) DownloadImage() error {
 	for _, vm := range c.Machines {
-		vm.DownloadImage()
+		err := vm.DownloadImage()
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
 
 func (c *Cluster) CreateDisks() error {
 	for _, vm := range c.Machines {
-		vm.CreateDisks()
+		err := vm.CreateDisks()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *Cluster) Create() error {
+	for _, vm := range c.Machines {
+		err := vm.Create()
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
 
 func (c *Cluster) Wait() error {
 	for _, vm := range c.Machines {
-		vm.Wait()
+		err := vm.Wait()
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
 
-func (c *Cluster) Shell() error {
+func (c *Cluster) Start() error {
 	for _, vm := range c.Machines {
-		vm.Shell()
+		err := vm.Start()
+		if err != nil {
+			return err
+		}
 	}
+	return nil
+}
+
+func (c *Cluster) Stop() error {
+	for _, vm := range c.Machines {
+		err := vm.Stop()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *Cluster) ForceStop() error {
+	for _, vm := range c.Machines {
+		err := vm.ForceStop()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *Cluster) Status() (string, error) {
+
+	return "", nil
+}
+
+func (c *Cluster) Delete() error {
+	for _, vm := range c.Machines {
+		err := vm.Delete()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *Cluster) CopyContent(origin string, dest string) error {
+
 	return nil
 }
 
 func (c *Cluster) RunInitScripts() error {
 	for _, vm := range c.Machines {
-		vm.RunInitScripts()
+		err := vm.RunInitScripts()
+		if err != nil {
+			return err
+		}
 	}
+	return nil
+}
+
+func (c *Cluster) Shell() error {
 	return nil
 }
